prov/efa: split efa_mr.h from efa.h
split MR related definitions from efa.h to efa_mr.h

Also rename efa_ep_is_hmem/cuda/neuron_mr to
efa_mr_is_hmem/cuda/neuron to conform with naming convention
