+++ prov/efa/src/rxr/rxr_atomic.c
@@ -137,7 +137,7 @@ ssize_t rxr_atomic_generic_efa(struct rxr_ep *rxr_ep,
 	assert(msg->iov_count <= rxr_ep->tx_iov_limit);
 	efa_perfset_start(rxr_ep, perf_efa_tx);
 
-	if (msg->desc && efa_ep_is_cuda_mr(msg->desc[0]))
+	if (msg->desc && efa_mr_is_cuda(msg->desc[0]))
 		return -FI_ENOSYS;
 
 	ofi_mutex_lock(&rxr_ep->util_ep.lock);
