+++ prov/efa/src/rxr/rxr_pkt_type_misc.c
@@ -359,7 +359,7 @@ void rxr_pkt_handle_readrsp_sent(struct rxr_ep *ep, struct rxr_pkt_entry *pkt_en
 		 * We currently require EFA RDMA support to enable the FI_HMEM
 		 * devices we support. Emulated read should not be used.
 		 */
-		assert(!efa_ep_is_hmem_mr(tx_entry->desc[0]));
+		assert(!efa_mr_is_hmem(tx_entry->desc[0]));
 
 		if (tx_entry->desc[0] || efa_is_cache_available(efa_domain))
 			rxr_prepare_desc_send(rxr_ep_domain(ep), tx_entry);
